+++ packages/opa/src/lib/config.ts
@@ -52,7 +52,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
     language: "typescript",
     openapiDocVersion: "0.2.0",
-    sdkVersion: "1.2.4",
-    genVersion: "2.361.10",
-    userAgent: "speakeasy-sdk/typescript 1.2.4 2.361.10 0.2.0 @styra/opa",
+    sdkVersion: "1.2.5",
+    genVersion: "2.365.0",
+    userAgent: "speakeasy-sdk/typescript 1.2.5 2.365.0 0.2.0 @styra/opa",
 } as const;