+++ packages/opa/RELEASES.md
@@ -390,3 +390,11 @@ Based on:
 - Speakeasy CLI 1.330.0 (2.361.10) https://github.com/speakeasy-api/speakeasy
 ### Generated
 - [typescript v1.2.4] packages/opa
+
+## 2024-07-11 00:20:41
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.331.0 (2.365.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.2.5] packages/opa